requirements.txt
pandas
sklearn
datetime
finnhub-python
flask
joblib
dash
requests
time